telegram_bot/bot.py
import telegram
from telegram.ext import Updater, CommandHandler, MessageHandler, Filters, CallbackContext
from telegram import Update

# Define your Telegram bot token
TELEGRAM_TOKEN = 'YOUR_TELEGRAM_BOT_TOKEN'

def start(update: Update, context: CallbackContext) -> None:
    update.message.reply_text('Welcome to the Real Estate Advisor Bot!')

def help_command(update: Update, context: CallbackContext) -> None:
    update.message.reply_text('This bot can help you with real estate queries. Use /start to begin.')

def handle_message(update: Update, context: CallbackContext) -> None:
    user_message = update.message.text
    # Here you can add logic to process the user's message and respond accordingly
    update.message.reply_text(f'You said: {user_message}')

def main() -> None:
    updater = Updater(TELEGRAM_TOKEN)

    dispatcher = updater.dispatcher

    dispatcher.add_handler(CommandHandler("start", start))
    dispatcher.add_handler(CommandHandler("help", help_command))
    dispatcher.add_handler(MessageHandler(Filters.text & ~Filters.command, handle_message))

    updater.start_polling()
    updater.idle()

if __name__ == '__main__':
    main()