requirements.txt
# Python dependencies
pandas
numpy
scikit-learn
xgboost
catboost
streamlit
python-telegram-bot
openai
jupyter
pytest